+++ Riot/Managers/Theme/Themes/DarkTheme.swift
@@ -30,8 +30,8 @@ class DarkTheme: NSObject, Theme {
     var searchBackgroundColor: UIColor = UIColor(rgb: 0x181B21)
     var searchPlaceholderColor: UIColor = UIColor(rgb: 0x61708B)
 
-    var headerBackgroundColor: UIColor = UIColor(rgb: 0x15171B)
-    var headerBorderColor: UIColor  = UIColor(rgb: 0x22262E)
+    var headerBackgroundColor: UIColor = UIColor(rgb: 0x22262E)
+    var headerBorderColor: UIColor  = UIColor(rgb: 0x181B21)
     var headerTextPrimaryColor: UIColor = UIColor(rgb: 0xA1B2D1)
     var headerTextSecondaryColor: UIColor = UIColor(rgb: 0xC8C8CD)
 
@@ -39,6 +39,7 @@ class DarkTheme: NSObject, Theme {
     var textSecondaryColor: UIColor = UIColor(rgb: 0xA1B2D1)
 
     var tintColor: UIColor = UIColor(rgb: 0x03B381)
+    var tintBackgroundColor: UIColor = UIColor(rgb: 0x1F6954)
     var unreadRoomIndentColor: UIColor = UIColor(rgb: 0x2E3648)
     var lineBreakColor: UIColor = UIColor(rgb: 0x61708B)