+++ Riot/Modules/Room/CellData/RoomBubbleCellData.h
@@ -35,6 +35,15 @@ typedef NS_ENUM(NSInteger, RoomBubbleCellDataTag)
  */
 @property(nonatomic) BOOL containsLastMessage;
 
+/**
+ Indicate true to display the timestamp of the selected component.
+ */
+@property(nonatomic) BOOL showTimestampForSelectedComponent;
+
+/**
+ Indicate true to display the timestamp of the selected component on the left if possible (YES by default).
+ */
+@property(nonatomic) BOOL displayTimestampForSelectedComponentOnLeftWhenPossible;
 
 /**
  The event id of the current selected event inside the bubble. Default is nil.