+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes in 0.8.7 (2019-xx-xx)
 ===============================================
 
+Improvements:
+ * RoomVC: When replying, use a "Reply" button instead of "Send".
+ * RoomVC: New message actions (#2394).
+ * Reactions: Display existing reactions below the message (#2396).
+ * Menu actions: Display message time (#2463).
+
 Bug fix:
  * Device Verification: Fix user display name and device id colors in dark theme